test user endpoint with rest-client
