+++ Gemfile
@@ -16,6 +16,7 @@ gem 'jbuilder', '~> 2.7'
 # Use Active Model has_secure_password
 gem 'bcrypt', '~> 3.1.7'
 gem "knock"
+gem 'email_validator'
 
 # Use Active Storage variant
 # gem 'image_processing', '~> 1.2'
