+++ app/controllers/api/v1/user_token_controller.rb
@@ -1,0 +1,7 @@
+module Api
+    module V1
+        class UserTokenController < Knock::AuthTokenController
+            skip_before_action :verify_authenticity_token, raise: false
+        end
+    end
+end
