+++ app/controllers/homepage_controller.rb
@@ -1,0 +1,4 @@
+class HomepageController < ApplicationController
+  def home
+  end
+end
