--- app/controllers/user_token_controller.rb
@@ -1,3 +1,0 @@
-class UserTokenController < Knock::AuthTokenController
-    skip_before_action :verify_authenticity_token, raise: false
-end
