+++ app/models/user.rb
@@ -1,5 +1,5 @@
 class User < ApplicationRecord
  has_secure_password
  validates :username, presence: true,  length: { in: 3..200 }
- validates :email, presence: true,uniqueness: true
+ validates :email, email: true
 end
