+++ app/views/api/v1/users/index.json.jbuilder
@@ -1,0 +1,1 @@
+json.partial! partial: 'user', collection: @users, as: :user
