+++ app/views/api/v1/users/show.json.jbuilder
@@ -1,0 +1,2 @@
+
+json.partial! partial: 'user', user: @user
