--- app/views/users/_user.json.jbuilder
@@ -1,2 +1,0 @@
-json.extract! user, :id, :username, :email, :password_digest, :firstname, :lastname, :isAdmin, :image, :created_at, :updated_at
-json.url user_url(user, format: :json)
